requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.0
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==1.0.0
Django==4.0.6
django-environ==0.9.0
django-social-share==2.2.1
gunicorn==20.1.0
idna==3.3
psycopg2-binary==2.9.3
requests==2.28.1
scraperapi-sdk==0.2.2
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
whitenoise==6.2.0
